feat: add authBar
